new statements : text, table, image
