+++ lotemplate/classes.py
@@ -12,19 +12,15 @@ __all__ = (
 
 import os
 from typing import Union
-from urllib import request
-from PIL import Image
 from sorcery import dict_of
 
 import uno
-import re
 import unohelper
-from com.sun.star.beans import PropertyValue, UnknownPropertyException
+from com.sun.star.beans import PropertyValue
 from com.sun.star.io import IOException
 from com.sun.star.lang import IllegalArgumentException, DisposedException
 from com.sun.star.connection import NoConnectException
 from com.sun.star.uno import RuntimeException
-from com.sun.star.awt import Size
 from com.sun.star.text.ControlCharacter import PARAGRAPH_BREAK
 from com.sun.star.style.BreakType import PAGE_AFTER
 
@@ -34,7 +30,9 @@ from .utils import *
 from lotemplate.Statement.ForStatement import ForStatement
 from lotemplate.Statement.HtmlStatement import HtmlStatement
 from lotemplate.Statement.IfStatement import IfStatement
-
+from lotemplate.Statement.TextStatement import TextStatement
+from lotemplate.Statement.TableStatement import TableStatement
+from lotemplate.Statement.ImageStatement import ImageStatement
 
 class Connexion:
 
@@ -194,109 +192,11 @@ class Template:
 
         should_close = kwargs.get("should_close", False)
 
-        def scan_text(doc) -> dict[str, dict[str, str]]:
-            """
-            scan for text in the given doc
-
-            :param doc: the document to scan
-            :return: the scanned variables
-            """
-
-            raw_string = doc.getText().getString()
-
-            matches = var_regexes['text'].finditer(raw_string)
-            plain_vars = {}
-            for var in matches:
-                key_name = var[0][1:]
-                # add to plain_vars if it doesn't matche ForStatement.foritem_regex
-                if not re.search(ForStatement.forindex_regex, key_name, re.IGNORECASE):
-                    plain_vars[key_name] = {'type': 'text', 'value': ''}
-
-            text_fields_vars = {}
-            for page in doc.getDrawPages():
-                for shape in page:
-                    try:
-                        matches = var_regexes['text'].finditer(shape.String)
-                    except (AttributeError, UnknownPropertyException):
-                        continue
-                    text_fields_vars = (text_fields_vars |
-                                        {var.group(0)[1:]: {'type': 'text', 'value': ''} for var in matches})
-
-            for var in scan_table(doc, get_list=True):
-                if '$' + var in plain_vars:
-                    del plain_vars[var]
-
-            for var in ForStatement.scan_for(doc):
-                if var in plain_vars:
-                    del plain_vars[var]
-
-            return plain_vars | text_fields_vars
-
-
-
-        def scan_table(doc, get_list=False) -> Union[dict, list]:
-            """
-            scan for tables in the given doc
-
-            :param get_list: indicates if the function should return a list
-            of variables or the formatted dictionary of variables
-            :param doc: the document to scan
-            :return: the scanned variables
-            """
-
-            def scan_cell(cell) -> None:
-                """
-                scan for variables in the given cell
-
-                :param cell: the cell to scan
-                :return: None
-                """
-                for match in var_regexes['table'].finditer(cell):
-                    if not match.captures('var'):
-                        continue
-                    if row_i != nb_rows - 1:
-                        raise errors.TemplateError(
-                            'variable_not_in_last_row',
-                            f"The variable {match[0]!r} (table {t_name!r}) "
-                            f"isn't in the last row (got: row {row_i + 1!r}, "
-                            f"expected: row {nb_rows!r})",
-                            dict(table=t_name, actual_row=row_i + 1,
-                                 expected_row=nb_rows, variable=match[0])
-                        )
-                    tab_vars[match[0][1:]] = {'type': 'table', 'value': ['']}
-                    list_tab_vars.append(match[0])
-
-            tab_vars = {}
-            list_tab_vars = []
-            for i in range(doc.getTextTables().getCount()):
-                table_data = doc.getTextTables().getByIndex(i).getDataArray()
-                t_name = doc.getTextTables().getByIndex(i).getName()
-                nb_rows = len(table_data)
-                for row_i, row in enumerate(table_data):
-                    for column in row:
-                        scan_cell(column)
-
-            return list_tab_vars if get_list else tab_vars
-
-        def scan_image(doc) -> dict[str, dict[str, str]]:
-            """
-            scan for images in the given doc
-
-            :param doc: the document to scan
-            :return: the scanned variables
-            """
-
-            return {
-                elem.LinkDisplayName[1:]: {'type': 'image', 'value': ''}
-                for elem in doc.getGraphicObjects()
-                if var_regexes['image'].fullmatch(elem.LinkDisplayName)
-            }
-
-        texts = scan_text(self.doc)
+        texts = TextStatement.scan_text(self.doc)
         # we use another document for if statement scanning because it modifies the file
         IfStatement.scan_if(template = self)
-        tables = scan_table(self.doc)
-        images = scan_image(self.doc)
+        tables = TableStatement.scan_table(self.doc)
+        images = ImageStatement.scan_image(self.doc)
         fors = ForStatement.scan_for(self.doc)
         HtmlStatement.scan_html(self.doc)
 
@@ -363,143 +263,6 @@ class Template:
         :return: None
         """
 
-
-        def text_fill(doc, variable: str, value: str) -> None:
-            """
-            Fills all the text-related content
-
-            :param doc: the document to fill
-            :param variable: the variable to search
-            :param value: the value to replace with
-            :return: None
-            """
-
-            search = doc.createSearchDescriptor()
-            search.SearchString = variable
-            founded = doc.findAll(search)
-
-            for x_found in founded:
-                text = x_found.getText()
-                cursor = text.createTextCursorByRange(x_found)
-                cursor.String = value
-
-            for page in doc.getDrawPages():
-                for shape in page:
-                    if shape.getShapeType() == "com.sun.star.drawing.TextShape":
-                        shape.String = shape.String.replace(variable, value)
-
-        def html_fill(doc, variable: str, value: str) -> None:
-            """
-            Fills all the html-related content
-
-            :param doc: the document to fill
-            :param variable: the variable to search
-            :param value: the value to replace with
-            :return: None
-            """
-
-            search = doc.createSearchDescriptor()
-            search.SearchString = variable
-            founded = doc.findAll(search)
-            for x_found in founded:
-                text = x_found.getText()
-                cursor = text.createTextCursorByRange(x_found)
-                cursor.String = ""
-                self.pasteHtml(value, cursor)
-
-            for page in doc.getDrawPages():
-                for shape in page:
-                    if shape.getShapeType() == "com.sun.star.drawing.TextShape":
-                        shape.String = shape.String.replace(variable, value)
-                        # we wanted to use the pasteHtml function, but it doesn't work in a shape
-                        # cursor = shape.createTextCursor()
-                        # oldString = cursor.String
-                        # self.pasteHtml(oldString.replace(variable, value), cursor)
-
-        def image_fill(doc, graphic_provider, variable: str, path: str, should_resize=True) -> None:
-            """
-            Fills all the image-related content
-
-            :param should_resize: specify if the image should be resized to keep his original size ratio
-            :param graphic_provider: the graphic provider, from the established connection
-            :param doc: the document to fill
-            :param variable: the variable to search
-            :param path: the path of the image to replace with
-            :return: None
-            """
-
-            if not path:
-                return
-
-            for graphic_object in doc.getGraphicObjects():
-                if graphic_object.LinkDisplayName != variable:
-                    continue
-
-                new_image = graphic_provider.queryGraphic((PropertyValue('URL', 0, get_file_url(path), 0),))
-
-                if should_resize:
-                    with Image.open(request.urlopen(path) if is_network_based(path) else path) as image:
-                        ratio = image.width / image.height
-                    new_size = Size()
-                    new_size.Height = graphic_object.Size.Height
-                    new_size.Width = graphic_object.Size.Height * ratio
-                    graphic_object.setSize(new_size)
-
-                graphic_object.Graphic = new_image
-
-        def tables_fill(doc, text_prefix: str, table_prefix: str) -> None:
-            """
-            Fills all the table-related content
-
-            :param doc: the document to fill
-            :param text_prefix: the prefix for text variables
-            :param table_prefix: the prefix for table variables
-            :return: None
-            """
-
-            search = doc.createSearchDescriptor()
-            matches = []
-            for element, infos in sorted(variables.items(), key=lambda s: -len(s[0])):
-                if infos['type'] != 'table':
-                    continue
-                search.SearchString = (text_prefix if '(' in element else table_prefix) + element
-                founded = doc.findAll(search)
-                matches += [founded.getByIndex(i) for i in range(founded.getCount()) if founded.getByIndex(i).TextTable]
-            tab_vars = [{
-                "table": variable.TextTable,
-                "var": variable.String
-            } for variable in matches]
-
-            tables = [
-                {'table': tab, 'vars':
-                    {tab_var['var']: variables[tab_var['var'][1:]]['value']
-                     for tab_var in tab_vars if tab_var['table'] == tab}
-                 } for tab in list(set(variable['table'] for variable in tab_vars))
-            ]
-
-            for element in tables:
-
-                table = element['table']
-                table_vars = element['vars']
-                var_row_pos = len(table.getRows()) - 1
-                nb_rows_to_add = max([len(variable) for variable in table_vars.values()])
-                table.getRows().insertByIndex(var_row_pos + 1, nb_rows_to_add - 1)
-                table_values = table.getDataArray()
-                var_row = table_values[var_row_pos]
-                static_rows = table_values[:var_row_pos]
-
-                for i in range(nb_rows_to_add):
-                    new_row = var_row
-                    for variable_name, variable_value in sorted(table_vars.items(), key=lambda s: -len(s[0])):
-                        new_row = tuple(
-                            elem.replace(
-                                variable_name, variable_value[i]
-                                if i < len(variable_value) else ""
-                            ) for elem in new_row
-                        )
-                    static_rows += (new_row,)
-                table.setDataArray(static_rows)
-
         if self.new:
             self.new.dispose()
             self.new.close(True)
@@ -532,15 +295,15 @@ class Template:
 
         for var, details in sorted(variables.items(), key=lambda s: -len(s[0])):
             if details['type'] == 'text':
-                text_fill(self.new, "$" + var, details['value'])
+                TextStatement.text_fill(self.new, "$" + var, details['value'])
             elif details['type'] == 'image':
-                image_fill(self.new, self.cnx.graphic_provider, "$" + var, details['value'])
+                ImageStatement.image_fill(self.new, self.cnx.graphic_provider, "$" + var, details['value'])
             elif details['type'] == 'html':
-                html_fill(self.new, "$" + var, details['value'])
+                HtmlStatement.html_fill(template=self, doc=self.new, variable="$" + var, value=details['value'])
 
         HtmlStatement.html_replace(template=self, doc=self.new)
 
-        tables_fill(self.new, '$', '&')
+        TableStatement.tables_fill(self.new, variables, '$', '&')
 
     def export(self, name: str, should_replace=False) -> Union[str, None]:
         """
